+++ RELEASE_NOTES.md
@@ -1,6 +1,10 @@
 Release notes
 ==============
 
+### Unreleased
+
+- Fixed option unwrapping not applying to `Async<_>` fields
+
 ### 0.1.1 (2024-09-23)
 
 - Fixed F# nullability not applying to non-paging parameters on connection-returning fields